NODE-312: Moved Servers back to api package.
